+++ package.json
@@ -24,6 +24,7 @@
     "react-bootstrap-carousel": "^4.1.1",
     "react-bootstrap-validation": "^0.1.11",
     "react-dom": "^17.0.2",
+    "react-joyride": "^2.3.2",
     "react-redux": "^7.2.5",
     "react-router-dom": "^6.0.1",
     "react-scripts": "4.0.3",
