+++ src/app/steps.js
@@ -1,0 +1,9 @@
+const steps = [
+  {
+    attachTo: { element: '#tour-start', on: 'auto' },
+    title: 'Welcome to MED3.0',
+    text: 'This is a guided tour of our site. You can quit this tour at any time by clicking exit.',
+  },
+];
+
+export default steps;
