+++ src/home/LandingPage.js
@@ -4,8 +4,7 @@ import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
 import Carousel from 'react-bootstrap/Carousel';
-import { ShepherdTour, ShepherdTourContext } from 'react-shepherd';
-import newSteps from './steps';
+import { ShepherdTourContext } from 'react-shepherd';
 import Header from '../common/Header';
 import image from '../images/landing_img.jpg';
 import {
@@ -60,26 +59,9 @@ const caroCaptionStyle = {
   bottom: 'auto',
 };
 
-const tourOptions = {
-  defaultStepOptions: {
-    cancelIcon: {
-      enabled: true,
-    },
-  },
-  useModalOverlay: true,
-};
-
-function ShepherdButton() {
+const LandingPage = () => {
   const tour = useContext(ShepherdTourContext);
 
-  return (
-    <button className="button dark" onClick={tour.start} type="button">
-      Start Tour
-    </button>
-  );
-}
-
-const LandingPage = () => {
   return (
     <div style={pageStyle}>
       <Header />
@@ -89,9 +71,6 @@ const LandingPage = () => {
         {/* <Box component="div" sx={{ visibility: 'hidden' }}>
           Visibility Hidden
         </Box> */}
-        <ShepherdTour steps={newSteps} tourOptions={tourOptions}>
-          <ShepherdButton />
-        </ShepherdTour>
         <Typography sx={{ flexGrow: 1, fontWeight: 'bold', marginBottom: '15px' }} color="black" variant="h6">{ probPrompt }</Typography>
         <Carousel variant="dark">
           <Carousel.Item>
@@ -155,6 +134,7 @@ const LandingPage = () => {
             </Carousel.Caption>
           </Carousel.Item>
         </Carousel>
+        <Button id="tour-start" onClick={tour.start}>Start tour</Button>
         <Box sx={{
           // justifyContent: 'center',
           display: 'grid',
@@ -163,7 +143,6 @@ const LandingPage = () => {
           marginTop: 2,
         }}
         >
-
           <Link sx={{ justifySelf: 'center' }} to="signup">
             <Button
               type="button"
