--- src/home/steps.js
@@ -1,50 +1,0 @@
-const steps = [
-  {
-    id: 'intro',
-    attachTo: { element: '.first-element', on: 'bottom' },
-    beforeShowPromise() {
-      return new Promise((resolve) => {
-        setTimeout(() => {
-          window.scrollTo(0, 0);
-          resolve();
-        }, 500);
-      });
-    },
-    buttons: [
-      {
-        classes: 'shepherd-button-secondary',
-        text: 'Exit',
-        type: 'cancel',
-      },
-      {
-        classes: 'shepherd-button-primary',
-        text: 'Back',
-        type: 'back',
-      },
-      {
-        classes: 'shepherd-button-primary',
-        text: 'Next',
-        type: 'next',
-      },
-    ],
-    classes: 'custom-class-name-1 custom-class-name-2',
-    highlightClass: 'highlight',
-    scrollTo: false,
-    cancelIcon: {
-      enabled: true,
-    },
-    title: 'Welcome to React-Shepherd!',
-    text: ['React-Shepherd is a JavaScript library for guiding users through your React app.'],
-    when: {
-      show: () => {
-        console.log('show step');
-      },
-      hide: () => {
-        console.log('hide step');
-      },
-    },
-  },
-  // ...
-];
-
-export default steps;
